Fixed bug in admin area
When logged in as admin, you could not see the images in your Media library. A small check was added to not execute any URL replacement if you are logged in as Administrator. If you are logged in as Admin and would like to see the URL replacement, open your WP installation in another browser.